magic-school/school/magic_school/user/api/serializers.py
from django.contrib.auth import get_user_model
from rest_framework import serializers
from user.models import Message
User = get_user_model()


class UserSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)
    class Meta:
        model = User
        fields = ["username","password"]

        # extra_kwargs = {
        #     "url": {"view_name": "user_api:user-detail", "lookup_field": "username"}
        # }

class MessageSerializer(serializers.ModelSerializer):
    """For Serializing Message"""
    sender = serializers.SlugRelatedField(many=False, slug_field='username', queryset=User.objects.all())
    receiver = serializers.SlugRelatedField(many=False, slug_field='username', queryset=User.objects.all())
    class Meta:
        model = Message
        fields = ['sender', 'receiver', 'message', 'timestamp']