magic-school/school/requirements/base.txt
# Django 
Django==4.0.4
django-kn-defaults==0.3.1
djvue==0.2.1
sentry-sdk==1.5.1
elastic-apm==6.7.2







# for windows please add following lib before installing compressor
# rcssmin==1.1.0 --install-option="--without-c-extensions"
django-compressor==3.1


djangorestframework==3.13.1
django-environ==0.8.1




